+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.4.1 (2025-04-09)
 
 * Bugfix to `highlight`
+* `watch` should build first
 
 ### 0.4.0 (2025-05-09)